backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geopandas==0.14.1
shapely==2.0.2
firebase-admin==6.2.0
python-jwt==3.3.3
requests==2.31.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
scipy==1.11.3
numpy==1.26.1
python-multipart==0.0.6
gunicorn==21.2.0
